+++ src/aks-preview/HISTORY.rst
@@ -12,6 +12,22 @@ To release a new version, please select a new version number (usually plus 1 to 
 Pending
 +++++++
 
+18.0.0b10
++++++++
+* Wrap the ARG call in the managed namespace list command
+
+18.0.0b9
++++++++
+* Add `--max-blocked-nodes` to the `az aks nodepool add/update/upgrade` commands.
+
+18.0.0b8
++++++++
+* Populate location of managed namespaces using managed cluster's location
+
+18.0.0b7
++++++++
+* Add option `--disable-http-proxy` to `az aks update`.
+
 18.0.0b6
 +++++++
 * Quality improvements for `az aks extension` and `az aks extension type` command groups