+++ src/scvmm/HISTORY.rst
@@ -2,6 +2,8 @@
 
 Release History
 ===============
+1.2.0
+* Add command `az scvmm vm create-from-machines`.
 
 1.1.2
 ++++++